Modified files with the NGUI in hopes of making it network friendly. Some stuff still not working correctly. Second DUI doesn't recognize commands for example

+++ Unity/Assets/Plugins/Network/CNetworkStream.cs
@@ -197,7 +197,7 @@ public class CNetworkStream
 	
 	public void Write(float _fValue)
 	{
-		m_cBitStream.Write((float)_fValue);
+		m_cBitStream.Write(_fValue);
 	}
